+++ source/renderer/app/components/widgets/forms/WalletsDropdown.js
@@ -40,6 +40,7 @@ type Props = {
   numberOfStakePools: number,
   wallets: Array<$Shape<Wallet>>,
   getStakePoolById: Function,
+  syncingLabel?: string,
 };
 
 type WalletOption = {
@@ -48,6 +49,8 @@ type WalletOption = {
   numberOfStakePools: number,
   detail: string,
   value: string,
+  syncing?: boolean,
+  syncingLabel?: string,
 };
 
 export default class WalletsDropdown extends Component<Props> {
@@ -58,13 +61,17 @@ export default class WalletsDropdown extends Component<Props> {
       numberOfStakePools,
       delegatedStakePool,
       isHardwareWallet,
+      syncing,
+      syncingLabel,
     }: WalletOption) => (
       <WalletsDropdownOption
+        isSyncing={syncing}
         label={label}
         numberOfStakePools={numberOfStakePools}
         detail={detail}
         delegatedStakePool={delegatedStakePool}
         isHardwareWallet={isHardwareWallet}
+        syncingLabel={syncingLabel}
       />
     ),
     selectionRenderer: ({
@@ -73,14 +80,18 @@ export default class WalletsDropdown extends Component<Props> {
       numberOfStakePools,
       delegatedStakePool,
       isHardwareWallet,
+      syncing,
+      syncingLabel,
     }: WalletOption) => (
       <WalletsDropdownOption
         selected
+        isSyncing={syncing}
         label={label}
         numberOfStakePools={numberOfStakePools}
         detail={detail}
         delegatedStakePool={delegatedStakePool}
         isHardwareWallet={isHardwareWallet}
+        syncingLabel={syncingLabel}
       />
     ),
     skin: SelectSkin,
@@ -114,14 +125,16 @@ export default class WalletsDropdown extends Component<Props> {
           currentStakePoolId = lastDelegationStakePoolId;
         }
         const delegatedStakePool = getStakePoolById(currentStakePoolId);
-        const detail = !isRestoring ? formattedWalletAmount(amount) : '-';
+        const detail = !isRestoring ? formattedWalletAmount(amount) : null;
         return {
           detail,
+          syncing: isRestoring,
           label,
           value,
           numberOfStakePools,
           delegatedStakePool,
           isHardwareWallet,
+          syncingLabel: this.props.syncingLabel,
         };
       }
     );
@@ -132,7 +145,7 @@ export default class WalletsDropdown extends Component<Props> {
     const selectOptions = omit({ ...props, topError }, 'options');
     return (
       <>
-        <Select options={walletsData} {...selectOptions} optionHeight={62} />
+        <Select options={walletsData} {...selectOptions} optionHeight={50} />
         {bottomError && <div className={styles.error}>{bottomError}</div>}
       </>
     );