+++ source/renderer/app/components/widgets/forms/WalletsDropdownOption.js
@@ -6,28 +6,73 @@ import { getColorFromRange } from '../../../utils/colors';
 import styles from './WalletsDropdownOption.scss';
 import StakePool from '../../../domains/StakePool';
 import hardwareWalletsIcon from '../../../assets/images/hardware-wallet/connect-ic.inline.svg';
+import LoadingSpinner from '../LoadingSpinner';
 
 export type WalletOption = {
   delegatedStakePool?: ?StakePool,
-  detail: string,
+  detail?: string,
   label: string,
   numberOfStakePools?: number,
   selected?: boolean,
+  isSyncing?: boolean,
+  syncingLabel?: string,
 };
 
 export default class WalletsDropdownOption extends Component<WalletOption> {
+  renderLabel = (label: string, isHardwareWallet: boolean) => {
+    return (
+      <div className={styles.label}>
+        {label}
+        {isHardwareWallet && (
+          <SVGInline
+            svg={hardwareWalletsIcon}
+            className={styles.hardwareWalletsIcon}
+          />
+        )}
+      </div>
+    );
+  };
+
+  renderLabelSyncing = (label: string, syncingLabel: string, isHardwareWallet: boolean) => {
+    return (
+      <div className={styles.label}>
+        {label}
+        {isHardwareWallet && (
+          <SVGInline
+            svg={hardwareWalletsIcon}
+            className={styles.hardwareWalletsIcon}
+          />
+        )}
+        <span className={styles.labelSync}> {syncingLabel}</span>
+      </div>
+    );
+  };
+
   renderLabelAndTicker = () => {
-    const { delegatedStakePool, label, numberOfStakePools, isHardwareWallet } = this.props;
+    const {
+      delegatedStakePool,
+      label,
+      numberOfStakePools,
+      isSyncing,
+      syncingLabel,
+      isHardwareWallet,
+    } = this.props;
     if (!delegatedStakePool || !numberOfStakePools) {
       return (
-        <div className={styles.label}>
-          {label}
-          {isHardwareWallet && (
-            <SVGInline
-              svg={hardwareWalletsIcon}
-              className={styles.hardwareWalletsIcon}
-            />
-          )}
+        <div className={styles.topRow}>
+          <div className={styles.topRowTicker}>
+            {isSyncing && syncingLabel
+              ? this.renderLabelSyncing(label, syncingLabel, isHardwareWallet)
+              : this.renderLabel(label, isHardwareWallet)
+            }
+          </div>
+          <div className={styles.topRowSync}>
+            {isSyncing && (
+              <div className={styles.syncing}>
+                <LoadingSpinner />
+              </div>
+            )}
+          </div>
         </div>
       );
     }
@@ -36,17 +81,19 @@ export default class WalletsDropdownOption extends Component<WalletOption> {
     const color = getColorFromRange(ranking, numberOfStakePools);
     return (
       <div className={styles.topRow}>
-        <div style={{ color }} className={styles.ticker}>
-          [{ticker}]
-        </div>
-        <div className={styles.label}>
-          {label}
-          {isHardwareWallet && (
-            <SVGInline
-              svg={hardwareWalletsIcon}
-              className={styles.hardwareWalletsIcon}
-            />
-          )}
+        <div className={styles.topRowTicker}>
+          <div style={{ color }} className={styles.ticker}>
+            [{ticker}]
+          </div>
+          <div className={styles.label}>
+            {label}
+            {isHardwareWallet && (
+              <SVGInline
+                svg={hardwareWalletsIcon}
+                className={styles.hardwareWalletsIcon}
+              />
+            )}
+          </div>
         </div>
       </div>
     );
@@ -60,7 +107,7 @@ export default class WalletsDropdownOption extends Component<WalletOption> {
     return (
       <div className={componentStyles}>
         {this.renderLabelAndTicker()}
-        <div className={styles.detail}>{detail}</div>
+        {detail && <div className={styles.detail}>{detail}</div>}
       </div>
     );
   }